feat: ImagesController 분리
이미지 불러오는 컨트롤러 따로 분리
